Move audio data to scriptable object
