+++ Assets/ScriptableObjects/Utils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ba5a5063283a1549bc6e9e1d0b78a97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
