+++ Assets/ScriptableObjects/Utils/AudioData.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d002bb28841c47847b102d35f12b90f8, type: 3}
+  m_Name: AudioData
+  m_EditorClassIdentifier: 
+  attackSound: {fileID: 8300000, guid: 2a1147a933b26134187ba49ae799bf9b, type: 3}
+  healthPickUpSound: {fileID: 8300000, guid: da73e66433b2bec458d5b29d46a06238, type: 3}
+  levelUpSound: {fileID: 8300000, guid: dc7deadb65931bb4784a65fbf0bbe349, type: 3}
+  defeatSound: {fileID: 8300000, guid: c40494920d3d20046885c2c2a2713428, type: 3}
+  takeDamageSound: {fileID: 8300000, guid: 0244cb1d187ac2c4fbcc3d7a4d088a18, type: 3}
