+++ Assets/ScriptableObjects/Utils/AudioData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 676ffa180feff934eb9b602c79e0236e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
