+++ Assets/ScriptableObjects/Utils/AudioDataSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d002bb28841c47847b102d35f12b90f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
